+++ CHANGELOG.md
@@ -147,3 +147,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [2.0.7] - 2022-03-25
 - Feature : View Job List - to list all jobs with its pool name
 - Feature : Support comments in enginelist file
+- Feature : Bugfix - Unable to pull vsam connector info during restore of backup